backend/requirements.txt
fastapi==0.102.0
uvicorn[standard]==0.23.1
torch>=2.1.0
transformers>=4.52.0
Pillow>=10.0.0
python-multipart>=0.0.6
google-generativeai
pillow
python-multipart